pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repoherald"
version = "0.1.0"
description = "Promote your GitHub repos on Reddit — AI-powered announcements"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "RepoHerald Contributors"}
]
keywords = ["reddit", "github", "promotion", "cli", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "praw>=7.7",
    "requests>=2.31",
    "openai>=1.0",
    "anthropic>=0.25",
    "google-generativeai>=0.5",
    "ollama>=0.2",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[project.scripts]
repoherald = "repoherald.cli:main"

[project.urls]
Homepage = "https://github.com/your-username/repoherald"
Issues = "https://github.com/your-username/repoherald/issues"

[tool.setuptools.packages.find]
include = ["repoherald*"]

[tool.ruff]
line-length = 100
target-version = "py310"
